Flatten auth package
