+++ auth/bcrypt.go
@@ -1,0 +1,17 @@
+package auth
+
+import "golang.org/x/crypto/bcrypt"
+
+type bcryptPasswordHash []byte
+
+func (h bcryptPasswordHash) MatchesPlaintext(plaintext string) bool {
+	return bcrypt.CompareHashAndPassword(h.Bytes(), []byte(plaintext)) == nil
+}
+
+func (h bcryptPasswordHash) String() string {
+	return string(h)
+}
+
+func (h bcryptPasswordHash) Bytes() []byte {
+	return []byte(h)
+}
