+++ auth/password_hash.go
@@ -1,0 +1,23 @@
+package auth
+
+import (
+	"golang.org/x/crypto/bcrypt"
+)
+
+type PasswordHash interface {
+	MatchesPlaintext(string) bool
+	String() string
+	Bytes() []byte
+}
+
+func NewPasswordHash(plaintext string) (PasswordHash, error) {
+	bytes, err := bcrypt.GenerateFromPassword([]byte(plaintext), bcrypt.DefaultCost)
+	if err != nil {
+		return nil, err
+	}
+	return bcryptPasswordHash(bytes), nil
+}
+
+func NewPasswordHashFromBytes(bytes []byte) PasswordHash {
+	return bcryptPasswordHash(bytes)
+}
